+++ android/app/src/main/kotlin/deckers/thibault/aves/channel/streams/StorageAccessStreamHandler.kt
@@ -21,7 +21,6 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.launch
-import java.io.FileOutputStream
 
 // starting activity to give access with the native dialog
 // breaks the regular `MethodChannel` so we use a stream channel instead
@@ -124,10 +123,8 @@ class StorageAccessStreamHandler(private val activity: Activity, arguments: Any?
         MainActivity.pendingStorageAccessResultHandlers[MainActivity.CREATE_FILE_REQUEST] = PendingStorageAccessResultHandler(null, { uri ->
             ioScope.launch {
                 try {
-                    activity.contentResolver.openOutputStream(uri)?.use { output ->
-                        output as FileOutputStream
-                        // truncate is necessary when overwriting a longer file
-                        output.channel.truncate(0)
+                    // truncate is necessary when overwriting a longer file
+                    activity.contentResolver.openOutputStream(uri, "wt")?.use { output ->
                         output.write(bytes)
                     }
                     success(true)