+++ lib/model/actions/entry_info_actions.dart
@@ -13,19 +13,24 @@ enum EntryInfoAction {
   // GeoTIFF
   showGeoTiffOnMap,
   // motion photo
+  convertMotionPhotoToStillImage,
   viewMotionPhotoVideo,
   // debug
   debug,
 }
 
 class EntryInfoActions {
-  static const all = [
+  static const common = [
     EntryInfoAction.editDate,
     EntryInfoAction.editLocation,
     EntryInfoAction.editRating,
     EntryInfoAction.editTags,
     EntryInfoAction.removeMetadata,
+  ];
+
+  static const formatSpecific = [
     EntryInfoAction.showGeoTiffOnMap,
+    EntryInfoAction.convertMotionPhotoToStillImage,
     EntryInfoAction.viewMotionPhotoVideo,
   ];
 }
@@ -48,6 +53,8 @@ extension ExtraEntryInfoAction on EntryInfoAction {
       case EntryInfoAction.showGeoTiffOnMap:
         return context.l10n.entryActionShowGeoTiffOnMap;
       // motion photo
+      case EntryInfoAction.convertMotionPhotoToStillImage:
+        return context.l10n.entryActionConvertMotionPhotoToStillImage;
       case EntryInfoAction.viewMotionPhotoVideo:
         return context.l10n.entryActionViewMotionPhotoVideo;
       // debug
@@ -87,8 +94,10 @@ extension ExtraEntryInfoAction on EntryInfoAction {
       case EntryInfoAction.showGeoTiffOnMap:
         return AIcons.map;
       // motion photo
+      case EntryInfoAction.convertMotionPhotoToStillImage:
+        return AIcons.convertToStillImage;
       case EntryInfoAction.viewMotionPhotoVideo:
-        return AIcons.motionPhoto;
+        return AIcons.openVideo;
       // debug
       case EntryInfoAction.debug:
         return AIcons.debug;